requirements.txt
Flask==3.1.0
flask-cors==5.0.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.4.4
python-dotenv==0.21.0
